pyproject.toml
[tool.poetry]
name = "face-emoji-streamlit"
version = "0.1.0"
description = ""
authors = ["alivelimb"]

[tool.poetry.dependencies]
python = "^3.9"
numpy = "^1.22.4"
streamlit = "^1.9.0"
augly = {extras = ["image"], version = "^1.0.0"}
face-recognition = "^1.3.0"
Pillow = "8.4.0"

[tool.poetry.dev-dependencies]
mypy = "^0.950"
flake8 = "^4.0.1"
black = "^22.3.0"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
